rename spans as logical steps, remove extra param from console phase, un-parametrize tests, merge test files to keep with covention
