+++ .github/CODEOWNERS
@@ -7,7 +7,7 @@
 src/snowflake/cli/api/metrics.py @snowflakedb/snowcli @snowflakedb/nade
 src/snowflake/cli/api/project/definition_conversion.py @snowflakedb/snowcli @snowflakedb/nade
 src/snowflake/cli/api/utils/ @snowflakedb/snowcli @snowflakedb/nade
-tests/api/metrics/ @snowflakedb/snowcli @snowflakedb/nade
+tests/api/test_metrics.py @snowflakedb/snowcli @snowflakedb/nade
 tests_common/__init__.py @snowflakedb/snowcli @snowflakedb/nade
 tests_common/conftest.py @snowflakedb/snowcli @snowflakedb/nade
 tests_common/deflake.py @snowflakedb/snowcli @snowflakedb/nade
