+++ src/snowflake/cli/_plugins/nativeapp/entities/application_package.py
@@ -43,7 +43,7 @@ from snowflake.cli._plugins.nativeapp.utils import needs_confirmation
 from snowflake.cli._plugins.stage.diff import DiffResult
 from snowflake.cli._plugins.stage.manager import StageManager
 from snowflake.cli._plugins.workspace.context import ActionContext
-from snowflake.cli.api.cli_global_context import start_cli_metrics_span
+from snowflake.cli.api.cli_global_context import span
 from snowflake.cli.api.entities.common import EntityBase, get_sql_executor
 from snowflake.cli.api.entities.utils import (
     drop_generic_object,
@@ -219,6 +219,7 @@ class ApplicationPackageEntity(EntityBase[ApplicationPackageEntityModel]):
             force=force,
         )
 
+    @span("drop_app_package")
     def action_drop(self, action_ctx: ActionContext, force_drop: bool, *args, **kwargs):
         console = self._workspace_ctx.console
         sql_executor = get_sql_executor()
@@ -549,6 +550,7 @@ class ApplicationPackageEntity(EntityBase[ApplicationPackageEntityModel]):
         compiler.compile_artifacts()
         return bundle_map
 
+    @span("deploy_app_package")
     def _deploy(
         self,
         bundle_map: BundleMap | None,
@@ -916,7 +918,7 @@ class ApplicationPackageEntity(EntityBase[ApplicationPackageEntityModel]):
             if validation_result["status"] == "FAIL":
                 raise SetupScriptFailedValidation()
 
-    @start_cli_metrics_span("get_validation_result")
+    @span("validate_setup_script")
     def get_validation_result(
         self, use_scratch_stage: bool, interactive: bool, force: bool
     ):
