+++ src/snowflake/cli/_plugins/workspace/manager.py
@@ -61,10 +61,7 @@ class WorkspaceManager:
             action_ctx = ActionContext(
                 get_entity=self.get_entity,
             )
-            with get_cli_context().metrics.start_span(
-                f"{type(entity).__name__}.{action.value}"
-            ):
-                return entity.perform(action, action_ctx, *args, **kwargs)
+            return entity.perform(action, action_ctx, *args, **kwargs)
         else:
             raise ValueError(f'This entity type does not support "{action.value}"')
