+++ tests/api/test_metrics.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import uuid
 from itertools import count
 from unittest import mock
@@ -21,6 +22,92 @@ from snowflake.cli.api.metrics import (
     CLIMetricsInvalidUsageError,
     CLIMetricsSpan,
 )
+
+
+def test_metrics_no_counters():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+
+    # then
+    assert metrics.counters == {}
+    assert metrics.get_counter("counter1") is None
+
+
+def test_metrics_set_one_counter():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+    metrics.set_counter("counter1", 1)
+
+    # then
+    assert metrics.counters == {"counter1": 1}
+    assert metrics.get_counter("counter1") == 1
+
+
+def test_metrics_increment_new_counter():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+    metrics.increment_counter("counter1")
+
+    # then
+    assert metrics.counters == {"counter1": 1}
+    assert metrics.get_counter("counter1") == 1
+
+
+def test_metrics_increment_existing_counter():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+    metrics.set_counter("counter1", 2)
+    metrics.increment_counter(name="counter1", value=2)
+
+    # then
+    assert metrics.counters == {"counter1": 4}
+    assert metrics.get_counter("counter1") == 4
+
+
+def test_metrics_set_multiple_counters():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+    metrics.set_counter("counter1", 1)
+    metrics.set_counter("counter2", 0)
+    metrics.set_counter(name="counter2", value=2)
+
+    # then
+    assert metrics.counters == {"counter1": 1, "counter2": 2}
+    assert metrics.get_counter("counter1") == 1
+    assert metrics.get_counter("counter2") == 2
+
+
+def test_metrics_set_default_new_counter():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+    metrics.set_counter_default("c1", 3)
+
+    # then
+    assert metrics.counters == {"c1": 3}
+
+
+def test_metrics_set_default_existing_counter():
+    # given
+    metrics = CLIMetrics()
+
+    # when
+    metrics.set_counter("c2", 2)
+    metrics.set_counter_default("c2", 1)
+
+    # then
+    assert metrics.counters == {"c2": 2}
 
 
 # we need to mock time.monotonic because on windows it does not
@@ -39,7 +126,7 @@ def create_spans(metrics: CLIMetrics, width: int, depth: int):
         if num_spans <= 0:
             return
 
-        with metrics.start_span(f"span-{next(counter)}"):
+        with metrics.span(f"span-{next(counter)}"):
             create_span(num_spans - 1)
 
     for _ in range(width):
@@ -64,7 +151,7 @@ def test_metrics_spans_single_span_no_error_or_parent(mock_time_monotonic):
     metrics = CLIMetrics()
 
     # when
-    with metrics.start_span("span1") as span1:
+    with metrics.span("span1") as span1:
         assert metrics.current_span is span1
 
     assert metrics.current_span is None
@@ -92,7 +179,7 @@ def test_metrics_spans_finish_early_is_idempotent():
     metrics = CLIMetrics()
 
     # when
-    with metrics.start_span("span1") as span1:
+    with metrics.span("span1") as span1:
         start_time = span1.start_time
         span1.finish()
         execution_time = span1.execution_time
@@ -109,10 +196,10 @@ def test_metrics_spans_parent_with_one_child(mock_time_monotonic):
     metrics = CLIMetrics()
 
     # when
-    with metrics.start_span("parent") as parent:
+    with metrics.span("parent") as parent:
         assert metrics.current_span is parent
 
-        with metrics.start_span("child") as child:
+        with metrics.span("child") as child:
             assert metrics.current_span is child
 
         assert metrics.current_span is parent
@@ -162,15 +249,15 @@ def test_metrics_spans_parent_with_two_children_same_name(mock_time_monotonic):
     metrics = CLIMetrics()
 
     # when
-    with metrics.start_span("parent") as parent:
+    with metrics.span("parent") as parent:
         assert metrics.current_span is parent
 
-        with metrics.start_span("child") as child1:
+        with metrics.span("child") as child1:
             assert metrics.current_span is child1
 
         assert metrics.current_span is parent
 
-        with metrics.start_span("child") as child2:
+        with metrics.span("child") as child2:
             assert metrics.current_span is child2
 
         assert metrics.current_span is parent
@@ -249,7 +336,7 @@ def test_metrics_spans_error_is_propagated():
 
     # when
     with pytest.raises(RuntimeError):
-        with metrics.start_span("step1"):
+        with metrics.span("step1"):
             raise RuntimeError()
 
     # then
@@ -264,7 +351,7 @@ def test_metrics_spans_empty_name_raises_error():
 
     # when
     with pytest.raises(CLIMetricsInvalidUsageError) as err:
-        with metrics.start_span(""):
+        with metrics.span(""):
             pass
 
     # then
